Game now playable
All the parts of the game have been hooked together and the main even
loop can now collect a user's guess.

I've also added support for command line args with CLAP. The user can
no configure the number of wring guesses they are allowed to make in
the game. They can also specify a file which can be used as a source
for the secret.